refactor(agent-host): extract the write-free turn preparation stage
submitMessage had grown into a 231-line, fifteen-step pipeline that every
upcoming capability (Skills, Office tools) would extend in place, and
runTurn threaded nine positional parameters. The stretch between
admission and the first durable write now lives in prepareTurn(), a
standalone planning stage over explicit ports that returns one frozen
TurnPlan and is testable without a Host instance; attachment admission
and materialization move to turnAttachments so the resources primitives
stay protocol-free. The Host keeps admission, reservation, execution,
and terminal settlement — submitMessage is 96 lines and runTurn takes
(sessionId, plan, state). No behavior change; the full agent suite is
green.
